requirements.txt
paho.mqtt
pyyaml
RPi.GPIO
